all fixed by Kirill answer
