+++ Backend/src/P2Project.Application/Volunteers/Delete/DeleteCommand.cs
@@ -1,0 +1,4 @@
+﻿namespace P2Project.Application.Volunteers.Delete
+{
+    public record DeleteCommand(Guid VolunteerId);
+}
