+++ Backend/src/P2Project.Application/Volunteers/Delete/DeleteHandler.cs
@@ -1,21 +1,10 @@
 ﻿using CSharpFunctionalExtensions;
-using FluentValidation;
 using Microsoft.Extensions.Logging;
 using P2Project.Domain.Shared;
 using P2Project.Domain.Shared.IDs;
 
 namespace P2Project.Application.Volunteers.Delete
 {
-    public record DeleteRequest(Guid VolunteerId);
-    public record DeleteCommand(Guid VolunteerId);
-    public class DeleteValidator :
-        AbstractValidator<DeleteRequest>
-    {
-        public DeleteValidator()
-        {
-            RuleFor(d => d.VolunteerId).NotEmpty();
-        }
-    }
     public class DeleteHandler
     {
         private readonly IVolunteersRepository _volunteersRepository;
@@ -28,18 +17,16 @@ namespace P2Project.Application.Volunteers.Delete
             _logger = logger;
         }
         public async Task<Result<Guid, Error>> Handle(
-            DeleteRequest request,
+            DeleteCommand command,
             CancellationToken cancellationToken = default)
         {
             var volunteerId = VolunteerId.CreateVolunteerId(
-                request.VolunteerId);
+                command.VolunteerId);
 
             var volunteerResult = await _volunteersRepository.GetById(
                 volunteerId, cancellationToken);
             if(volunteerResult.IsFailure)
-                return Errors.General.NotFound(request.VolunteerId);
-
-            var command = new DeleteCommand(volunteerId);
+                return Errors.General.NotFound(command.VolunteerId);
 
             var id = await _volunteersRepository.Delete(
                 volunteerResult.Value, cancellationToken);
