+++ Backend/src/P2Project.Application/Volunteers/Delete/DeleteRequest.cs
@@ -1,0 +1,4 @@
+﻿namespace P2Project.Application.Volunteers.Delete
+{
+    public record DeleteRequest(Guid VolunteerId);
+}
