+++ Backend/src/P2Project.Application/Volunteers/Delete/DeleteValidator.cs
@@ -1,0 +1,13 @@
+﻿using FluentValidation;
+
+namespace P2Project.Application.Volunteers.Delete
+{
+    public class DeleteValidator :
+        AbstractValidator<DeleteRequest>
+    {
+        public DeleteValidator()
+        {
+            RuleFor(d => d.VolunteerId).NotEmpty();
+        }
+    }
+}
